requirements.txt
asgiref==3.2.3
Django==3.0
progressbar2==3.47.0
psycopg2==2.8.4
python-utils==2.3.0
pytz==2019.3
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
